Move profile resource to profile/description
