+++ app/views/dashboards/show.html.erb
@@ -8,7 +8,7 @@
   <div class="col-md order-md-last">
     <% if !current_user.home_location? %>
       <div id="map" class="content_map border border-secondary-subtle">
-        <p class="m-3"><%= t(".no_home_location_html", :edit_profile_link => link_to(t(".edit_your_profile"), profile_path)) %></p>
+        <p class="m-3"><%= t(".no_home_location_html", :edit_profile_link => link_to(t(".edit_your_profile"), profile_description_path)) %></p>
       </div>
     <% else %>
       <% content_for :head do %>
