+++ app/views/users/show.html.erb
@@ -286,6 +286,6 @@
 
 <% if current_user and @user.id == current_user.id %>
   <div class="my-3">
-    <%= link_to t(".edit_profile"), profile_path, :class => "btn btn-outline-primary" %>
+    <%= link_to t(".edit_profile"), profile_description_path, :class => "btn btn-outline-primary" %>
   </div>
 <% end %>
