+++ config/locales/en.yml
@@ -1994,33 +1994,34 @@ en:
         title: My Advanced Preferences
         save: Update Preferences
   profiles:
-    show:
-      title: Edit Profile
-      save: Update Profile
-      cancel: Cancel
-      image: Image
-      gravatar:
-        gravatar: "Use Gravatar"
-        link: "https://wiki.openstreetmap.org/wiki/Gravatar"
-        what_is_gravatar: "What is Gravatar?"
-      new image: "Add an image"
-      keep image: "Keep the current image"
-      delete image: "Remove the current image"
-      replace image: "Replace the current image"
-      image size hint: "(square images at least 100x100 work best)"
-      home location: "Home Location"
-      no home location: "You have not entered your home location."
-      update home location on click: "Update home location when I click on the map?"
-      show: "Show"
-      delete: "Delete"
-      undelete: "Undo delete"
-      social_links:
-        title: Social Profile Links
-        remove: Remove
-        add: Add Social Link
-    update:
-      success: Profile updated.
-      failure: Couldn't update profile.
+    descriptions:
+      show:
+        title: Edit Profile
+        save: Update Profile
+        cancel: Cancel
+        image: Image
+        gravatar:
+          gravatar: "Use Gravatar"
+          link: "https://wiki.openstreetmap.org/wiki/Gravatar"
+          what_is_gravatar: "What is Gravatar?"
+        new image: "Add an image"
+        keep image: "Keep the current image"
+        delete image: "Remove the current image"
+        replace image: "Replace the current image"
+        image size hint: "(square images at least 100x100 work best)"
+        home location: "Home Location"
+        no home location: "You have not entered your home location."
+        update home location on click: "Update home location when I click on the map?"
+        show: "Show"
+        delete: "Delete"
+        undelete: "Undo delete"
+        social_links:
+          title: Social Profile Links
+          remove: Remove
+          add: Add Social Link
+      update:
+        success: Profile updated.
+        failure: Couldn't update profile.
   sessions:
     new:
       tab_title: "Log In"
